+++ packages/nc-gui/composables/useColumnFilteredOrSorted.ts
@@ -1,0 +1,52 @@
+import { type FilterType } from 'nocodb-sdk'
+
+export function useColumnFilteredOrSorted() {
+  // TODO: handle nested filters
+  const { nestedFilters: _nestedFilters, allFilters, sorts } = useSmartsheetStoreOrThrow()
+  const filteredColumnIds = computed(() => {
+    const columnIds: Set<string> = new Set<string>()
+    const extractFilterArray = (filters: FilterType[]) => {
+      if (filters && filters.length > 0) {
+        for (const eachFilter of filters) {
+          if (eachFilter.is_group) {
+            if ((eachFilter.children?.length ?? 0) > 0) {
+              extractFilterArray(eachFilter.children!)
+            }
+          } else if (eachFilter.fk_column_id) {
+            columnIds.add(eachFilter.fk_column_id)
+          }
+        }
+      }
+    }
+    extractFilterArray(allFilters.value)
+    return columnIds
+  })
+  const sortedColumnIds = computed(() => {
+    const columnIds: Set<string> = new Set<string>()
+    if (!sorts?.value || sorts.value.length === 0) {
+      return columnIds
+    }
+    for (const sort of sorts.value) {
+      if (sort.fk_column_id) {
+        columnIds.add(sort.fk_column_id)
+      }
+    }
+    return columnIds
+  })
+
+  const isColumnSortedOrFiltered = (colId: string) => {
+    if (filteredColumnIds.value.has(colId)) {
+      return 'FILTERED'
+    } else if (sortedColumnIds.value.has(colId)) {
+      return 'SORTED'
+    } else {
+      return undefined
+    }
+  }
+
+  return {
+    filteredColumnIds,
+    sortedColumnIds,
+    isColumnSortedOrFiltered,
+  }
+}
